setToggleButtonState함수 js파일추가 분리후 모듈화
